+++ src/briefy/leica/models/job/location.py
@@ -31,7 +31,7 @@ class JobLocation(LocationContactInfoMixin, AddressMixin,
 
     __summary_attributes__ = [
         'id', 'country', 'locality', 'coordinates', 'email', 'mobile',
-        'additional_phone', 'fullname',
+        'additional_phone', 'fullname', 'formatted_address'
     ]
 
     __listing_attributes__ = __summary_attributes__